2019 version 0.1.6 add a dialog for self adapting samples generation in ui/sampleProduce
